angular-sample/angular-webapp/client/app/document/company-documents/components/personalized-contract-list/personalized-contract-list.component.spec.ts
import { RestClientServiceStub } from './../../../../shared/testing/mocks/rest-client-service-stub';
import { RestClientService } from './../../../../shared/data/rest-client.service';
import { EmployeeSearchService } from './../../../../employee/services/employee-search.service';
import { addOrUpdateAtlasParamValue } from '../../../../root-module/common/extract-helpers';
import { SortDirection } from '../../../../atlas-elements/common/models/ae-sort-model';
import { AtlasApiRequestWithParams, AtlasParams } from './../../../../shared/models/atlas-api-response';
import { AeSelectItem } from './../../../../atlas-elements/common/models/ae-select-item';
import { Observable } from 'rxjs/Rx';
import { LoadSitesCompleteAction, LoadSitesAction, LoadAllDepartmentsAction, LoadAllDepartmentsCompleteAction } from './../../../../shared/actions/company.actions';
import { async, ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { ReactiveFormsModule } from '@angular/forms';
import { By } from '@angular/platform-browser';
import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
import { ActivatedRoute, Router, RouterModule } from '@angular/router';
import { Store, StoreModule } from '@ngrx/store';
import { InjectorRef, LocaleService, LocalizationModule, TranslationService } from 'angular-l10n';
import { Document, DocumentActionType } from './../../../models/document';
import { AeAutocompleteComponent } from '../../../../atlas-elements/ae-autocomplete/ae-autocomplete.component';
import { AeDatatableComponent } from '../../../../atlas-elements/ae-datatable/ae-datatable.component';
import { AeInputComponent } from '../../../../atlas-elements/ae-input/ae-input.component';
import { AeNavActionsComponent } from '../../../../atlas-elements/ae-nav-actions/ae-nav-actions.component';
import { AtlasElementsModule } from '../../../../atlas-elements/atlas-elements.module';
import { BreadcrumbService } from '../../../../atlas-elements/common/services/breadcrumb-service';
import { HomeModule } from '../../../../home/home.module';
import { routes } from '../../../../home/home.routes';
import { AtlasSharedModule } from '../../../../shared/atlas-shared.module';
import { ClaimsHelperService } from '../../../../shared/helpers/claims-helper';
import * as fromRoot from '../../../../shared/reducers';
import { reducer } from '../../../../shared/reducers/index';
import { ActivatedRouteStub } from '../../../../shared/testing/mocks/activated-route-stub';
import { BreadcrumbServiceStub } from '../../../../shared/testing/mocks/breadcrumb-service-mock';
import { ClaimsHelperServiceStub } from '../../../../shared/testing/mocks/claims-helper-service-mock';
import { LocaleServiceStub } from '../../../../shared/testing/mocks/locale-service-stub';
import { MockStoreCompanyDocuments } from '../../../../shared/testing/mocks/mock-store-company-documents';
import { RouterMock } from '../../../../shared/testing/mocks/router-stub';
import { TranslationServiceStub } from '../../../../shared/testing/mocks/translation-service-stub';
import { DocumentSharedModule } from '../../../document-shared/document-shared.module';
import { ContractsListComponent } from '../contracts-list/contracts-list.component';
import { PersonalizedContractListComponent } from './personalized-contract-list.component';
import { LoadPersonalContractsListCompleteAction, LoadContractsListAction } from '../../actions/contracts.actions';
import { CommonTestHelper } from '../../../../shared/testing/helpers/common-test-helper';
import * as Immutable from 'immutable';

describe('Company documents -> contracts & hand books ->Personalised contracts', () => {
    let component: PersonalizedContractListComponent;
    let fixture: ComponentFixture<PersonalizedContractListComponent>;
    let store: Store<fromRoot.State>;
    let loadedSites: AeSelectItem<string>[];
    let loadedDepartments: AeSelectItem<string>[];
    let mockedAeselectSitesData: AeSelectItem<string>[];
    let mockedAeselectDepartmentData: AeSelectItem<string>[];
    let dispatchSpy: jasmine.Spy;
    let params: AtlasParams[] = new Array();
    params.push(new AtlasParams('contractsFilter', 2))
    let contractsApiRequest = new AtlasApiRequestWithParams(1, 10, 'Title', SortDirection.Ascending, params);
    let rows; let actions;
    let columns;

    beforeEach(async(() => {
        TestBed.configureTestingModule({
            imports: [
                LocalizationModule
                , ReactiveFormsModule
                , AtlasElementsModule
                , DocumentSharedModule
                , StoreModule.provideStore(reducer)
                , RouterModule.forChild(routes)
                , AtlasSharedModule
                , HomeModule
                , BrowserAnimationsModule
            ]
            , declarations: [
                PersonalizedContractListComponent
                , ContractsListComponent
            ]
            , providers: [
                InjectorRef
                , { provide: BreadcrumbService, useClass: BreadcrumbServiceStub }
                , { provide: LocaleService, useClass: LocaleServiceStub }
                , { provide: TranslationService, useClass: TranslationServiceStub }
                , { provide: ClaimsHelperService, useClass: ClaimsHelperServiceStub }
                , { provide: Router, useClass: RouterMock }
                , { provide: ActivatedRoute, useClass: ActivatedRouteStub }
                , EmployeeSearchService
                , { provide: RestClientService, useClass: RestClientServiceStub }
            ]
        })
            .overrideComponent(PersonalizedContractListComponent, { set: { host: { '(click)': 'dummy' } } })
            .compileComponents();
    }));

    describe('Loading personalised contracts list with out data', () => {
        beforeEach(() => {

            fixture = TestBed.createComponent(PersonalizedContractListComponent);
            component = fixture.componentInstance;
            store = fixture.debugElement.injector.get(Store);
            dispatchSpy = spyOn(store, 'dispatch');
            fixture.detectChanges();
            jasmine.DEFAULT_TIMEOUT_INTERVAL = 30000;

        });

        it('should create', () => {
            expect(component).toBeTruthy();
            let dataTable = fixture.debugElement.query(By.directive(AeDatatableComponent));
            expect(dataTable).toBeTruthy();
        });

        it('should have a filter by Site and have a filter by site of auto complete type', () => {
            let siteFilter = fixture.debugElement.query(By.css('#ddlSite'));
            expect(siteFilter).toBeTruthy();
            expect(siteFilter.componentInstance instanceof AeInputComponent).toBeTruthy();
            let autoCompleteElement = fixture.debugElement.query(By.directive(AeAutocompleteComponent));
            expect(autoCompleteElement).toBeTruthy();
            expect(autoCompleteElement.name).toEqual('ae-autocomplete');

        });
        it('should load sites,personalised contractList,departments when no sites,personalised contractList,departments found in the store', () => {
            let siteLoadActionPayLoad: LoadSitesAction = (new LoadSitesAction(false));
            expect(dispatchSpy).toHaveBeenCalledWith(siteLoadActionPayLoad);
            let contractLoadAction: LoadContractsListAction = (new LoadContractsListAction(component.personalizedApiRequest));
            expect(dispatchSpy).toHaveBeenCalledWith(contractLoadAction);
            let departmentLoadAction: LoadAllDepartmentsAction = (new LoadAllDepartmentsAction());
            expect(dispatchSpy).toHaveBeenCalledWith(departmentLoadAction);
        });


        describe('After site, document category, documents list is loaded', () => {
            beforeEach(fakeAsync(() => {
                let mockedSites = MockStoreCompanyDocuments.getSiteLists();
                let mockedDepartments = MockStoreCompanyDocuments.getDepartmentLists();
                mockedAeselectSitesData = MockStoreCompanyDocuments.getSiteAeelement();
                mockedAeselectDepartmentData = MockStoreCompanyDocuments.getDepartmentAeelement();
                component.sites$.subscribe((val) => {
                    loadedSites = val;
                });
                component.departments$.subscribe((val) => {
                    loadedDepartments = val;
                });

                dispatchSpy.and.callThrough();

                store.dispatch(new LoadSitesCompleteAction(mockedSites.Entities));
                store.dispatch(new LoadPersonalContractsListCompleteAction(MockStoreCompanyDocuments.getPersonalisedContract()));
                store.dispatch(new LoadAllDepartmentsCompleteAction(mockedDepartments.Entities));
                fixture.detectChanges();

                let siteControlDebug = fixture.debugElement.query(By.css('#ddlSite'));
                siteControlDebug.triggerEventHandler('change', { target: { value: null } });

                fixture.detectChanges();
                tick(100);
                rows = fixture.debugElement.queryAll(By.css('.table__row'));
                columns = fixture.debugElement.queryAll(By.css('.table__heading'));


            }));

            it('Load sites,departments for filtering', fakeAsync(() => {
                tick(100);
                expect(loadedSites).toEqual(mockedAeselectSitesData);
                expect(loadedDepartments).toEqual(mockedAeselectDepartmentData);
            }));


            it('should have a column for employee name,template,last modified date,actions with sort option', () => {
                expect(columns[0].nativeElement.innerText.trim()).toEqual('HANDBOOK.EMPLOYEE_NAME');
                expect(columns[0].nativeElement.classList).toContain('js-sortable');
                expect(columns[1].nativeElement.innerText.trim()).toEqual('HANDBOOK.LAST_MODIFY');
                expect(columns[1].nativeElement.classList).toContain('js-sortable');
                expect(columns[2].nativeElement.innerText.trim()).toEqual('HANDBOOK.TEMPLATE');
                expect(columns[2].nativeElement.classList).toContain('js-sortable');
                expect(columns[3].nativeElement.innerText.trim()).toEqual('HANDBOOK.DISTRIBUTED');
                expect(columns[3].nativeElement.classList).toContain('js-sortable');

                expect(columns[4].nativeElement.innerText.trim()).toEqual('Actions');
                expect(columns[4].nativeElement.classList).not.toContain('js-sortable');
            });

            it('should have 4 action items and have an action item as download as pdf,Distribute,Update,Save as pdf when a document is not distributed', fakeAsync(() => {
                let actionButtonitems = rows[9].query(By.css('.nav--actions')).componentInstance;
                expect(actionButtonitems instanceof AeNavActionsComponent).toBeTruthy();
                let actionButton = rows[9].query(By.css('.nav--actions')).componentInstance;
                let event = new MouseEvent('click');
                actionButton._onClick(event);
                tick(200);
                fixture.detectChanges();
                actions = rows[9].queryAll(By.css('li'));
                expect(actions.length).toEqual(4);
                //Below cases are failing while running in server so using other methodology to assert these case
                // expect(actions[0].nativeElement.innerText.toString().toLowerCase()).toEqual('view');
                // expect(actions[1].nativeElement.innerText.toString().toLowerCase()).toEqual('download pdf');
                // expect(actions[2].nativeElement.innerText.toString().toLowerCase()).toEqual('distribute');
                // expect(actions[3].nativeElement.innerText.toString().toLowerCase()).toEqual('update');
                // expect(actions[4].nativeElement.innerText.toString().toLowerCase()).toEqual('save as pdf');
                let listComponent = <ContractsListComponent>fixture.debugElement.query(By.directive(ContractsListComponent)).componentInstance;
                CommonTestHelper.hasGivenButton(listComponent.actions.toArray(), 'download pdf');
                CommonTestHelper.hasGivenButton(listComponent.actions.toArray(), 'distribute');
                CommonTestHelper.hasGivenButton(listComponent.actions.toArray(), 'update');
                CommonTestHelper.hasGivenButton(listComponent.actions.toArray(), 'save as pdf');
            }));

            it('should have 4 action items and have an action item as view,download as pdf,Update,Save as pdf when a document is distributed', fakeAsync(() => {
                let actionButtonitems = rows[0].query(By.css('.nav--actions')).componentInstance;
                expect(actionButtonitems instanceof AeNavActionsComponent).toBeTruthy();
                let actionButton = rows[0].query(By.css('.nav--actions')).componentInstance;
                let event = new MouseEvent('click');
                actionButton._onClick(event);
                tick(200);
                fixture.detectChanges();
                actions = rows[0].queryAll(By.css('li'));
                expect(actions.length).toEqual(4);
                let listComponent = <ContractsListComponent>fixture.debugElement.query(By.directive(ContractsListComponent)).componentInstance;
                CommonTestHelper.hasGivenButton(listComponent.actions.toArray(), 'view');
                CommonTestHelper.hasGivenButton(listComponent.actions.toArray(), 'download pdf');
                CommonTestHelper.hasGivenButton(listComponent.actions.toArray(), 'update');
                CommonTestHelper.hasGivenButton(listComponent.actions.toArray(), 'save as pdf');
            }));

            it('User selects sites filter', fakeAsync(() => {
                let currentRequest = component.personalizedApiRequest;
                let selectedSite = new AeSelectItem<string>('Site 1', '42bbe9d6-8730-29d2-7cf3-aa526e976bb5', false);
                component.personalcontractListForm.patchValue({ site: [selectedSite] });
                fixture.detectChanges();
                tick(200);
                currentRequest.Params = addOrUpdateAtlasParamValue(currentRequest.Params, 'site', '42bbe9d6-8730-29d2-7cf3-aa526e976bb5');
                let companyDocLoadAction: LoadContractsListAction = (new LoadContractsListAction(currentRequest));
                expect(dispatchSpy).toHaveBeenCalledWith(companyDocLoadAction);

                store.dispatch(new LoadPersonalContractsListCompleteAction(MockStoreCompanyDocuments.getFilterSitePC()));
                dispatchSpy.and.callThrough();
                fixture.detectChanges();
            }));
            it('User selects department filter', fakeAsync(() => {
                let currentRequest = component.personalizedApiRequest;
                let selectedDepartment = new AeSelectItem<string>('testteam', '7b83d7e2-49f2-45cb-8a75-73210ccc5fc8', false);
                component.personalcontractListForm.patchValue({ department: [selectedDepartment] });
                fixture.detectChanges();
                tick(400);
                currentRequest.Params = addOrUpdateAtlasParamValue(currentRequest.Params, 'department', '7b83d7e2-49f2-45cb-8a75-73210ccc5fc8');
                let companyDocLoadAction: LoadContractsListAction = (new LoadContractsListAction(currentRequest));
                expect(dispatchSpy).toHaveBeenCalledWith(companyDocLoadAction);

                store.dispatch(new LoadPersonalContractsListCompleteAction(MockStoreCompanyDocuments.getFilterDepartmentPC()));
                dispatchSpy.and.callThrough();
                fixture.detectChanges();
            })); //
            it('User selects employee filter', fakeAsync(() => {
                let currentRequest = component.personalizedApiRequest;
                let selectedEmp = new AeSelectItem<string>('Bruce  preemp', '69a710a4-2eb1-4155-82cb-0e6112a0ee61', false);
                component.personalcontractListForm.patchValue({ employee: [selectedEmp] });
                fixture.detectChanges();
                tick(600);
                currentRequest.Params = addOrUpdateAtlasParamValue(currentRequest.Params, 'employee', '69a710a4-2eb1-4155-82cb-0e6112a0ee61');
                let companyDocLoadAction: LoadContractsListAction = (new LoadContractsListAction(currentRequest));
                expect(dispatchSpy).toHaveBeenCalledWith(companyDocLoadAction);

                store.dispatch(new LoadPersonalContractsListCompleteAction(MockStoreCompanyDocuments.getFilterEmployeePC()));
                dispatchSpy.and.callThrough();
                fixture.detectChanges();
            }));

        });

    });
})